Split storage handler tests by protocol concern
