+++ crates/core/src/message/handlers/storage.rs
@@ -542,1661 +542,19 @@ impl HandleMsg<SyncEntriesWithSuccessorReport> for MessageHandler {
 
 #[cfg(not(feature = "wasm"))]
 #[cfg(test)]
-mod test {
-    use std::sync::Arc;
-
-    use tokio::time::timeout;
-    use tokio::time::Duration;
-
-    use super::*;
-    use crate::consts::ENTRY_DATA_MAX_LEN;
-    use crate::dht::entry::PlacedEntry;
-    use crate::dht::entry::PlacementMiss;
-    use crate::dht::successor::SuccessorReader;
-    use crate::dht::successor::SuccessorWriter;
-    use crate::dht::Chord;
-    use crate::ecc::tests::gen_ordered_keys;
-    use crate::ecc::SecretKey;
-    use crate::message::Encoder;
-    use crate::message::MessageRelay;
-    use crate::message::Transaction;
-    use crate::prelude::entry::EntryKind;
-    use crate::session::SessionSk;
-    use crate::storage::MemStorage;
-    use crate::swarm::callback::SwarmCallback;
-    use crate::swarm::transport::STORAGE_LOOKUP_OBSERVATION_CAPACITY;
-    use crate::swarm::SwarmBuilder;
-    use crate::tests::default::assert_no_more_msg;
-    use crate::tests::default::prepare_node;
-    use crate::tests::default::wait_for_msgs;
-    use crate::tests::default::Node;
-    use crate::tests::manually_establish_connection;
-
-    struct NoopCallback;
-
-    impl SwarmCallback for NoopCallback {}
-
-    async fn next_payload(node: &Node) -> Result<MessagePayload> {
-        node.listen_once()
-            .await
-            .ok_or_else(|| Error::InvalidMessage("expected message payload".to_string()))
-    }
-
-    async fn next_payload_for_tx(node: &Node, tx_id: uuid::Uuid) -> Result<MessagePayload> {
-        for _ in 0..64 {
-            let payload = timeout(Duration::from_secs(1), node.listen_once())
-                .await
-                .map_err(|_| {
-                    Error::InvalidMessage("timed out waiting for matching payload".to_string())
-                })?
-                .ok_or_else(|| Error::InvalidMessage("expected message payload".to_string()))?;
-            if payload.transaction.tx_id == tx_id {
-                return Ok(payload);
-            }
-        }
-
-        Err(Error::InvalidMessage(
-            "matching transaction payload was not observed".to_string(),
-        ))
-    }
-
-    fn next_generated_key(keys: &mut impl Iterator<Item = SecretKey>) -> Result<SecretKey> {
-        keys.next()
-            .ok_or_else(|| Error::InvalidMessage("expected generated key".to_string()))
-    }
-
-    fn storage_sync_report_payload(
-        request: &MessagePayload,
-        report: SyncEntriesWithSuccessorReport,
-        signer: &SessionSk,
-        next_hop: Did,
-        destination: Did,
-    ) -> Result<MessagePayload> {
-        let transaction = Transaction::new(
-            destination,
-            request.transaction.tx_id,
-            Message::SyncEntriesWithSuccessorReport(report),
-            signer,
-        )?;
-        let relay = MessageRelay::new(vec![signer.account_did()], next_hop, destination);
-        MessagePayload::new(transaction, signer, relay)
-    }
-
-    fn prepare_node_with_storage_redundancy(key: SecretKey, redundancy: u16) -> Result<Node> {
-        let session_sk = SessionSk::new_with_seckey(&key)?;
-        let swarm = Arc::new(
-            SwarmBuilder::new(
-                0,
-                "stun://stun.l.google.com:19302",
-                Box::new(MemStorage::new()),
-                session_sk,
-            )
-            .dht_storage_redundancy(redundancy)
-            .dht_finger_table_size(8)
-            .build(),
-        );
-        Ok(Node::new(swarm))
-    }
-
-    fn prepare_node_with_virtual_nodes(key: SecretKey, positions_per_peer: u16) -> Result<Node> {
-        let session_sk = SessionSk::new_with_seckey(&key)?;
-        let swarm = Arc::new(
-            SwarmBuilder::new(
-                0,
-                "stun://stun.l.google.com:19302",
-                Box::new(MemStorage::new()),
-                session_sk,
-            )
-            .dht_virtual_nodes(positions_per_peer)
-            .dht_finger_table_size(8)
-            .build(),
-        );
-        Ok(Node::new(swarm))
-    }
-
-    fn owner_index(nodes: &[&Node], placement: Did) -> Result<usize> {
-        let mut owner = None;
-        for (index, node) in nodes.iter().enumerate() {
-            if !matches!(
-                node.dht().find_successor(placement)?,
-                PeerRingAction::Some(_)
-            ) {
-                continue;
-            }
-
-            if owner.replace(index).is_some() {
-                return Err(Error::InvalidMessage(
-                    "placement has more than one observed owner".to_string(),
-                ));
-            }
-        }
-        owner.ok_or_else(|| Error::InvalidMessage("placement has no observed owner".to_string()))
-    }
-
-    fn physical_sync_route_next_hop(dht: &PeerRing, destination: Did) -> Result<Option<Did>> {
-        if destination == dht.did {
-            return Ok(None);
-        }
-
-        match dht.find_successor(destination)? {
-            PeerRingAction::Some(next) if next == dht.did => Ok(Some(destination)),
-            PeerRingAction::Some(next) => Ok(Some(next)),
-            PeerRingAction::RemoteAction(next, PeerRingRemoteAction::FindSuccessor(_)) => {
-                Ok(Some(next))
-            }
-            action => Err(Error::unexpected_peer_ring_action(action)),
-        }
-    }
-
-    fn storage_sync_route_next_hop(dht: &PeerRing, placement: Did) -> Result<Option<Did>> {
-        match dht.find_storage_owner(placement)? {
-            PeerRingAction::Some(_) => Ok(None),
-            PeerRingAction::RemoteAction(next, PeerRingRemoteAction::FindSuccessor(_)) => {
-                Ok(Some(next))
-            }
-            action => Err(Error::unexpected_peer_ring_action(action)),
-        }
-    }
-
-    fn remote_storage_placement_after(node: &Node, start: Did) -> Result<Did> {
-        for offset in 1..512 {
-            let placement = start + Did::from(offset);
-            if matches!(
-                node.dht().find_storage_owner(placement)?,
-                PeerRingAction::RemoteAction(_, PeerRingRemoteAction::FindSuccessor(key))
-                    if key == placement
-            ) {
-                return Ok(placement);
-            }
-        }
-
-        Err(Error::InvalidMessage(
-            "expected a remote storage placement".to_string(),
-        ))
-    }
-
-    fn install_two_node_chord_view(first: &Node, second: &Node) -> Result<()> {
-        first.dht().successors().update(second.did())?;
-        second.dht().successors().update(first.did())?;
-        *first.dht().lock_predecessor()? = Some(second.did());
-        *second.dht().lock_predecessor()? = Some(first.did());
-        Ok(())
-    }
-
-    fn split_redundant_entry(nodes: &[&Node]) -> Result<(Entry, Did, Did, usize, usize)> {
-        for attempt in 0..512 {
-            let topic = format!("split remote replica placement {attempt}");
-            let entry: Entry = topic.try_into()?;
-            let mut placements = entry.did.rotate_affine(2)?.into_iter();
-            let primary = placements
-                .next()
-                .ok_or_else(|| Error::InvalidMessage("expected primary placement".to_string()))?;
-            let replica = placements
-                .next()
-                .ok_or_else(|| Error::InvalidMessage("expected replica placement".to_string()))?;
-            let primary_owner = owner_index(nodes, primary)?;
-            let replica_owner = owner_index(nodes, replica)?;
-            if primary_owner != replica_owner {
-                return Ok((entry, primary, replica, primary_owner, replica_owner));
-            }
-        }
-
-        Err(Error::InvalidMessage(
-            "could not sample a split-owner redundant entry".to_string(),
-        ))
-    }
-
-    fn non_affine_placement(entry_key: Did, redundancy: u16) -> Result<Did> {
-        let placements = entry_key.rotate_affine(redundancy)?;
-        for attempt in 0..512 {
-            let candidate = Entry::gen_did(&format!("non-affine placement {attempt}"))?;
-            if !placements.contains(&candidate) {
-                return Ok(candidate);
-            }
-        }
-
-        Err(Error::InvalidMessage(
-            "could not sample non-affine placement".to_string(),
-        ))
-    }
-
-    async fn assert_cached_data_values(
-        node: &Node,
-        entry_key: Did,
-        expected: &[&str],
-    ) -> Result<()> {
-        let entry = node
-            .swarm
-            .storage_check_cache(entry_key)
-            .await
-            .ok_or_else(|| Error::InvalidMessage("expected cached entry".to_string()))?;
-        let expected_data = expected
-            .iter()
-            .map(|value| value.to_string().encode())
-            .collect::<Result<Vec<_>>>()?;
-
-        assert_eq!(entry.did, entry_key);
-        assert_eq!(entry.kind, EntryKind::Data);
-        assert_eq!(entry.data, expected_data);
-        assert_eq!(entry.crdt.dots.len(), entry.data.len());
-        Ok(())
-    }
-
-    #[test]
-    fn finish_storage_action_accepts_empty_action() -> Result<()> {
-        finish_storage_action(PeerRingAction::None)?;
-        Ok(())
-    }
-
-    #[test]
-    fn finish_storage_action_rejects_unhandled_action() -> Result<()> {
-        let did = SecretKey::random().address().into();
-        match finish_storage_action(PeerRingAction::Some(did)) {
-            Err(Error::PeerRingUnexpectedAction(action)) => {
-                assert_eq!(*action, PeerRingAction::Some(did));
-                Ok(())
-            }
-            res => Err(Error::InvalidMessage(format!(
-                "expected unexpected storage action, got {res:?}"
-            ))),
-        }
-    }
-
-    #[tokio::test]
-    async fn sync_entries_handler_stores_entry_at_placement_key() -> Result<()> {
-        let node = prepare_node(SecretKey::random()).await;
-        let handler = MessageHandler::new(node.swarm.transport.clone(), Arc::new(NoopCallback));
-        let resource_id = Did::from(10u32);
-        let placement_key = Did::from(100u32);
-        let entry = Entry::new(
-            resource_id,
-            vec!["placed".to_string().encode()?],
-            EntryKind::Data,
-        );
-        let stored_entry = entry.clone().try_into_storage_entry()?;
-        let context_key = SecretKey::random();
-        let context_session = SessionSk::new_with_seckey(&context_key)?;
-        let context = MessagePayload::new_send(
-            Message::custom(b"sync context")?,
-            &context_session,
-            node.did(),
-            node.did(),
-        )?;
-
-        handler
-            .handle(&context, &SyncEntriesWithSuccessor {
-                destination: StorageSyncDestination::PhysicalOwner(node.did()),
-                data: vec![PlacedEntry::new(placement_key, entry.clone())],
-            })
-            .await?;
-
-        assert_eq!(
-            node.dht().storage.get(&placement_key.to_string()).await?,
-            Some(stored_entry)
-        );
-        assert_eq!(
-            node.dht().storage.get(&resource_id.to_string()).await?,
-            None
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn sync_entries_handler_caps_inbound_entry_payloads() -> Result<()> {
-        let node = prepare_node(SecretKey::random()).await;
-        let handler = MessageHandler::new(node.swarm.transport.clone(), Arc::new(NoopCallback));
-        let placement_key = Did::from(100u32);
-        let entry = Entry::new(
-            Did::from(10u32),
-            (0..ENTRY_DATA_MAX_LEN + 3)
-                .map(|i| format!("payload{i}").encode())
-                .collect::<Result<Vec<_>>>()?,
-            EntryKind::Data,
-        );
-        let context_key = SecretKey::random();
-        let context_session = SessionSk::new_with_seckey(&context_key)?;
-        let context = MessagePayload::new_send(
-            Message::custom(b"sync context")?,
-            &context_session,
-            node.did(),
-            node.did(),
-        )?;
-
-        handler
-            .handle(&context, &SyncEntriesWithSuccessor {
-                destination: StorageSyncDestination::PhysicalOwner(node.did()),
-                data: vec![PlacedEntry::new(placement_key, entry)],
-            })
-            .await?;
-
-        let stored = node
-            .dht()
-            .storage
-            .get(&placement_key.to_string())
-            .await?
-            .ok_or_else(|| Error::InvalidMessage("expected stored sync entry".to_string()))?;
-
-        assert_eq!(stored.data.len(), ENTRY_DATA_MAX_LEN);
-        let first_payload: String = stored
-            .data
-            .first()
-            .ok_or_else(|| Error::InvalidMessage("expected capped payload".to_string()))?
-            .decode()?;
-        assert_eq!(first_payload, String::from("payload3"));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn sync_entries_handler_accepts_placement_destination_on_local_branch() -> Result<()> {
-        let mut keys = gen_ordered_keys(2).into_iter();
-        let node1 = prepare_node(next_generated_key(&mut keys)?).await;
-        let node2 = prepare_node(next_generated_key(&mut keys)?).await;
-        manually_establish_connection(&node1.swarm, &node2.swarm).await;
-        wait_for_msgs([&node1, &node2]).await;
-        assert_no_more_msg([&node1, &node2]).await;
-        install_two_node_chord_view(&node1, &node2)?;
-
-        let handler = MessageHandler::new(node1.swarm.transport.clone(), Arc::new(NoopCallback));
-        let placement_key = node2.did();
-        assert!(matches!(
-            node1.dht().find_storage_owner(placement_key)?,
-            PeerRingAction::Some(witness) if witness == node2.did() && witness != node1.did()
-        ));
-        let entry = Entry::new(
-            Did::from(10u32),
-            vec!["routed repair".to_string().encode()?],
-            EntryKind::Data,
-        );
-        let stored_entry = entry.clone().try_into_storage_entry()?;
-        let msg = SyncEntriesWithSuccessor {
-            destination: StorageSyncDestination::PlacementKey(placement_key),
-            data: vec![PlacedEntry::new(placement_key, entry.clone())],
-        };
-        let context = MessagePayload::new_send(
-            Message::SyncEntriesWithSuccessor(msg.clone()),
-            node2.swarm.transport.session_sk(),
-            node1.did(),
-            placement_key,
-        )?;
-
-        handler.handle(&context, &msg).await?;
-
-        assert_eq!(
-            node1.dht().storage.get(&placement_key.to_string()).await?,
-            Some(stored_entry.clone())
-        );
-        let ack = next_payload(&node2).await?;
-        assert!(matches!(
-            ack.transaction.data()?,
-            Message::SyncEntriesWithSuccessorReport(SyncEntriesWithSuccessorReport {
-                destination,
-                receiver,
-                acks
-            }) if destination == StorageSyncDestination::PlacementKey(placement_key)
-                && receiver == node1.did()
-                && acks == vec![SyncedEntryAck::new(placement_key, stored_entry)]
-        ));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn sync_entries_handler_rejects_mismatched_placement_destination() -> Result<()> {
-        let mut keys = gen_ordered_keys(2).into_iter();
-        let sender = prepare_node(next_generated_key(&mut keys)?).await;
-        let receiver = prepare_node(next_generated_key(&mut keys)?).await;
-        manually_establish_connection(&sender.swarm, &receiver.swarm).await;
-        wait_for_msgs([&sender, &receiver]).await;
-        assert_no_more_msg([&sender, &receiver]).await;
-        install_two_node_chord_view(&sender, &receiver)?;
-
-        let destination_key = receiver.did();
-        let mismatched_key = sender.did();
-        let entry = Entry::new(
-            Did::from(10u32),
-            vec!["mismatched placement".to_string().encode()?],
-            EntryKind::Data,
-        );
-        let sync_msg = SyncEntriesWithSuccessor {
-            destination: StorageSyncDestination::PlacementKey(destination_key),
-            data: vec![PlacedEntry::new(mismatched_key, entry)],
-        };
-        let context = MessagePayload::new_send(
-            Message::SyncEntriesWithSuccessor(sync_msg.clone()),
-            sender.swarm.transport.session_sk(),
-            receiver.did(),
-            destination_key,
-        )?;
-        let receiver_handler =
-            MessageHandler::new(receiver.swarm.transport.clone(), Arc::new(NoopCallback));
-
-        receiver_handler.handle(&context, &sync_msg).await?;
-
-        assert_eq!(
-            receiver
-                .dht()
-                .storage
-                .get(&mismatched_key.to_string())
-                .await?,
-            None
-        );
-        let payload = next_payload_for_tx(&sender, context.transaction.tx_id).await?;
-        assert!(matches!(
-            payload.transaction.data::<Message>()?,
-            Message::SyncEntriesWithSuccessorReport(SyncEntriesWithSuccessorReport { acks, .. })
-                if acks.is_empty()
-        ));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn sync_entries_handler_rejects_physical_destination_for_unowned_placement() -> Result<()>
-    {
-        let mut keys = gen_ordered_keys(2).into_iter();
-        let sender = prepare_node(next_generated_key(&mut keys)?).await;
-        let receiver = prepare_node(next_generated_key(&mut keys)?).await;
-        manually_establish_connection(&sender.swarm, &receiver.swarm).await;
-        wait_for_msgs([&sender, &receiver]).await;
-        assert_no_more_msg([&sender, &receiver]).await;
-        install_two_node_chord_view(&sender, &receiver)?;
-
-        let placement_key = remote_storage_placement_after(&receiver, sender.did())?;
-        let entry = Entry::new(
-            Did::from(10u32),
-            vec!["wrong physical owner".to_string().encode()?],
-            EntryKind::Data,
-        );
-        let sync_msg = SyncEntriesWithSuccessor {
-            destination: StorageSyncDestination::PhysicalOwner(receiver.did()),
-            data: vec![PlacedEntry::new(placement_key, entry)],
-        };
-        let context = MessagePayload::new_send(
-            Message::SyncEntriesWithSuccessor(sync_msg.clone()),
-            sender.swarm.transport.session_sk(),
-            receiver.did(),
-            receiver.did(),
-        )?;
-        let receiver_handler =
-            MessageHandler::new(receiver.swarm.transport.clone(), Arc::new(NoopCallback));
-
-        receiver_handler.handle(&context, &sync_msg).await?;
-
-        assert_eq!(
-            receiver
-                .dht()
-                .storage
-                .get(&placement_key.to_string())
-                .await?,
-            None
-        );
-        let payload = next_payload_for_tx(&sender, context.transaction.tx_id).await?;
-        assert!(matches!(
-            payload.transaction.data::<Message>()?,
-            Message::SyncEntriesWithSuccessorReport(SyncEntriesWithSuccessorReport { acks, .. })
-                if acks.is_empty()
-        ));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn sync_entries_handler_acks_local_branch_with_successor_witness() -> Result<()> {
-        let mut keys = gen_ordered_keys(2).into_iter();
-        let sender = prepare_node(next_generated_key(&mut keys)?).await;
-        let receiver = prepare_node(next_generated_key(&mut keys)?).await;
-        manually_establish_connection(&sender.swarm, &receiver.swarm).await;
-        wait_for_msgs([&sender, &receiver]).await;
-        assert_no_more_msg([&sender, &receiver]).await;
-        install_two_node_chord_view(&sender, &receiver)?;
-
-        let placement_key = sender.did();
-        assert!(matches!(
-            receiver.dht().find_storage_owner(placement_key)?,
-            PeerRingAction::Some(witness) if witness == sender.did() && witness != receiver.did()
-        ));
-
-        let entry = Entry::new(
-            Did::from(10u32),
-            vec!["successor witness owner".to_string().encode()?],
-            EntryKind::Data,
-        );
-        let stored_entry = entry.clone().try_into_storage_entry()?;
-        let sync_msg = SyncEntriesWithSuccessor {
-            destination: StorageSyncDestination::PhysicalOwner(receiver.did()),
-            data: vec![PlacedEntry::new(placement_key, entry)],
-        };
-        let context = MessagePayload::new_send(
-            Message::SyncEntriesWithSuccessor(sync_msg.clone()),
-            sender.swarm.transport.session_sk(),
-            receiver.did(),
-            receiver.did(),
-        )?;
-        let receiver_handler =
-            MessageHandler::new(receiver.swarm.transport.clone(), Arc::new(NoopCallback));
-
-        receiver_handler.handle(&context, &sync_msg).await?;
-
-        assert_eq!(
-            receiver
-                .dht()
-                .storage
-                .get(&placement_key.to_string())
-                .await?,
-            Some(stored_entry.clone())
-        );
-        let payload = next_payload_for_tx(&sender, context.transaction.tx_id).await?;
-        assert!(matches!(
-            payload.transaction.data::<Message>()?,
-            Message::SyncEntriesWithSuccessorReport(SyncEntriesWithSuccessorReport { acks, .. })
-                if acks == vec![SyncedEntryAck::new(placement_key, stored_entry)]
-        ));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn leave_dht_republishes_after_responsibility_peer_departure() -> Result<()> {
-        let key = SecretKey::random();
-        let session = SessionSk::new_with_seckey(&key)?;
-        let swarm = Arc::new(
-            SwarmBuilder::new(
-                0,
-                "stun://stun.l.google.com:19302",
-                Box::new(MemStorage::new()),
-                session,
-            )
-            .dht_storage_redundancy(2)
-            .build(),
-        );
-        let node = Node::new(swarm);
-        let departed = Did::from(100u32);
-        node.dht().successors().update(departed)?;
-        let entry = Entry::new(key.address().into(), vec![], EntryKind::Data);
-        let placement_keys = entry.did.rotate_affine(2)?;
-        node.dht()
-            .storage
-            .put(&placement_keys[0].to_string(), &entry)
-            .await?;
-        let handler = MessageHandler::new(node.swarm.transport.clone(), Arc::new(NoopCallback));
-
-        handler.leave_dht(departed).await?;
-
-        assert!(!node.dht().successors().contains(&departed)?);
-        assert_eq!(
-            node.dht()
-                .storage
-                .get(&placement_keys[1].to_string())
-                .await?,
-            Some(entry)
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn storage_api_rejects_redundancy_mismatch() -> Result<()> {
-        let node = prepare_node(SecretKey::random()).await;
-
-        let result =
-            <Swarm as ChordStorageInterface<2>>::storage_fetch(&node.swarm, node.did()).await;
-
-        assert!(matches!(
-            result,
-            Err(Error::StorageRedundancyMismatch {
-                configured: 1,
-                requested: 2
-            })
-        ));
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn placed_entry_operation_rejects_non_affine_placement() -> Result<()> {
-        let node = prepare_node_with_storage_redundancy(SecretKey::random(), 2)?;
-        let handler = MessageHandler::new(node.swarm.transport.clone(), Arc::new(NoopCallback));
-        let topic = "reject misplaced remote storage operation".to_string();
-        let entry: Entry = topic.try_into()?;
-        let invalid_placement = non_affine_placement(entry.did, 2)?;
-        let msg = PlacedEntryOperation {
-            placement: invalid_placement,
-            op: EntryOperation::Overwrite(entry.clone()),
-        };
-        let sender_session = SessionSk::new_with_seckey(&SecretKey::random())?;
-        let context = MessagePayload::new_send(
-            Message::OperateEntry(msg.clone()),
-            &sender_session,
-            node.did(),
-            node.did(),
-        )?;
-
-        assert!(!msg.placement_belongs_to_entry(2)?);
-        let result = handler.handle(&context, &msg).await;
-
-        assert!(matches!(
-            result,
-            Err(Error::InvalidMessage(message)) if message.contains("affine replica set")
-        ));
-        assert_eq!(
-            node.dht()
-                .storage
-                .get(&invalid_placement.to_string())
-                .await?,
-            None
-        );
-        assert_eq!(node.dht().storage.get(&entry.did.to_string()).await?, None);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn remote_redundant_store_writes_split_replica_at_affine_placement() -> Result<()> {
-        let mut keys = gen_ordered_keys(2).into_iter();
-        let node1 = prepare_node_with_storage_redundancy(next_generated_key(&mut keys)?, 2)?;
-        let node2 = prepare_node_with_storage_redundancy(next_generated_key(&mut keys)?, 2)?;
-
-        manually_establish_connection(&node1.swarm, &node2.swarm).await;
-        wait_for_msgs([&node1, &node2]).await;
-        assert_no_more_msg([&node1, &node2]).await;
-
-        let nodes = [&node1, &node2];
-        let (entry, primary, replica, primary_owner, replica_owner) =
-            split_redundant_entry(&nodes)?;
-        assert_eq!(primary, entry.did);
-        let writer = nodes[primary_owner];
-        let remote_replica_owner = nodes[replica_owner];
-
-        <Swarm as ChordStorageInterface<2>>::storage_store(&writer.swarm, entry.clone()).await?;
-
-        let payload = next_payload(remote_replica_owner).await?;
-        assert!(matches!(
-            payload.transaction.data()?,
-            Message::OperateEntry(PlacedEntryOperation {
-                placement,
-                op: EntryOperation::Overwrite(remote_entry),
-            }) if placement == replica && remote_entry.did == entry.did
-        ));
-        assert_eq!(
-            writer
-                .dht()
-                .storage
-                .get(&primary.to_string())
-                .await?
-                .map(|stored| stored.did),
-            Some(entry.did)
-        );
-        assert_eq!(
-            remote_replica_owner
-                .dht()
-                .storage
-                .get(&replica.to_string())
-                .await?
-                .map(|stored| stored.did),
-            Some(entry.did)
-        );
-        assert_eq!(
-            remote_replica_owner
-                .dht()
-                .storage
-                .get(&primary.to_string())
-                .await?,
-            None
-        );
-        assert_no_more_msg([&node1, &node2]).await;
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn local_hit_read_repair_sends_no_search_for_unknown_replicas() -> Result<()> {
-        let key = SecretKey::random();
-        let session = SessionSk::new_with_seckey(&key)?;
-        let swarm = Arc::new(
-            SwarmBuilder::new(
-                0,
-                "stun://stun.l.google.com:19302",
-                Box::new(MemStorage::new()),
-                session,
-            )
-            .dht_storage_redundancy(2)
-            .build(),
-        );
-        let node = Node::new(swarm);
-        let entry = Entry::new(
-            key.address().into(),
-            vec!["local".to_string().encode()?],
-            EntryKind::Data,
-        );
-        let first_key = entry
-            .did
-            .rotate_affine(2)?
-            .into_iter()
-            .next()
-            .ok_or_else(|| Error::InvalidMessage("expected first placement".to_string()))?;
-        node.dht()
-            .storage
-            .put(&first_key.to_string(), &entry)
-            .await?;
-
-        <Swarm as ChordStorageInterface<2>>::storage_fetch(&node.swarm, entry.did).await?;
-
-        assert_eq!(node.swarm.storage_check_cache(entry.did).await, Some(entry));
-        assert_no_more_msg([&node]).await;
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn found_entry_repairs_buffered_misses_only() -> Result<()> {
-        let node = prepare_node(SecretKey::random()).await;
-        let handler = MessageHandler::new(node.swarm.transport.clone(), Arc::new(NoopCallback));
-        let entry = Entry::new(
-            Did::from(10u32),
-            vec!["repair".to_string().encode()?],
-            EntryKind::Data,
-        );
-        let stored_entry = entry.clone().try_into_storage_entry()?;
-        let placement_key = Did::from(100u32);
-        let unknown_key = Did::from(120u32);
-        let context_key = SecretKey::random();
-        let context_session = SessionSk::new_with_seckey(&context_key)?;
-        let context = MessagePayload::new_send(
-            Message::FoundEntry(FoundEntry {
-                data: vec![],
-                misses: vec![PlacementMiss::new(placement_key, node.did())],
-                resource: entry.did,
-                redundancy: 2,
-            }),
-            &context_session,
-            node.did(),
-            node.did(),
-        )?;
-
-        handler
-            .handle(&context, &FoundEntry {
-                data: vec![],
-                misses: vec![PlacementMiss::new(placement_key, node.did())],
-                resource: entry.did,
-                redundancy: 2,
-            })
-            .await?;
-        handler
-            .handle(&context, &FoundEntry {
-                data: vec![entry.clone()],
-                misses: vec![],
-                resource: entry.did,
-                redundancy: 2,
-            })
-            .await?;
-
-        assert_eq!(
-            node.dht().storage.get(&placement_key.to_string()).await?,
-            Some(stored_entry)
-        );
-        assert_eq!(
-            node.dht().storage.get(&unknown_key.to_string()).await?,
-            None
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn found_entry_rejects_multiple_entries() -> Result<()> {
-        let node = prepare_node(SecretKey::random()).await;
-        let handler = MessageHandler::new(node.swarm.transport.clone(), Arc::new(NoopCallback));
-        let resource = Did::from(10u32);
-        let first = Entry::new(
-            resource,
-            vec!["first".to_string().encode()?],
-            EntryKind::Data,
-        );
-        let second = Entry::new(
-            resource,
-            vec!["second".to_string().encode()?],
-            EntryKind::Data,
-        );
-        let context_key = SecretKey::random();
-        let context_session = SessionSk::new_with_seckey(&context_key)?;
-        let context = MessagePayload::new_send(
-            Message::FoundEntry(FoundEntry {
-                data: vec![first.clone(), second.clone()],
-                misses: vec![],
-                resource,
-                redundancy: 2,
-            }),
-            &context_session,
-            node.did(),
-            node.did(),
-        )?;
-
-        let result = handler
-            .handle(&context, &FoundEntry {
-                data: vec![first, second],
-                misses: vec![],
-                resource,
-                redundancy: 2,
-            })
-            .await;
-
-        assert!(
-            matches!(result, Err(Error::InvalidMessage(message)) if message.contains("more than one"))
-        );
-        assert_eq!(node.swarm.storage_check_cache(resource).await, None);
-        assert_eq!(node.swarm.transport.storage_lookup_observation_count()?, 0);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn storage_miss_observation_buffer_is_bounded() -> Result<()> {
-        let node = prepare_node(SecretKey::random()).await;
-        for index in 0..(STORAGE_LOOKUP_OBSERVATION_CAPACITY + 8) {
-            let resource = Did::from((index + 1) as u32);
-            let placement = Did::from((index + 10_000) as u32);
-            node.swarm
-                .transport
-                .observe_storage_misses(resource, 2, [PlacementMiss::new(placement, node.did())])?;
-        }
-
-        assert!(
-            node.swarm.transport.storage_lookup_observation_count()?
-                <= STORAGE_LOOKUP_OBSERVATION_CAPACITY
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn storage_fetch_starts_fresh_observation_round() -> Result<()> {
-        let node = prepare_node(SecretKey::random()).await;
-        let resource = Did::from(10u32);
-        let placement = Did::from(100u32);
-        node.swarm
-            .transport
-            .observe_storage_misses(resource, 1, [PlacementMiss::new(placement, node.did())])?;
-
-        node.swarm.transport.start_storage_lookup(resource, 1)?;
-        let misses = node.swarm.transport.take_storage_misses(resource, 1)?;
-
-        assert!(misses.is_empty());
-        assert_eq!(node.swarm.transport.storage_lookup_observation_count()?, 0);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn expired_storage_misses_do_not_trigger_late_repair() -> Result<()> {
-        let node = prepare_node(SecretKey::random()).await;
-        let handler = MessageHandler::new(node.swarm.transport.clone(), Arc::new(NoopCallback));
-        let entry = Entry::new(
-            Did::from(10u32),
-            vec!["fresh".to_string().encode()?],
-            EntryKind::Data,
-        );
-        let placement_key = Did::from(100u32);
-        let context_key = SecretKey::random();
-        let context_session = SessionSk::new_with_seckey(&context_key)?;
-        let context = MessagePayload::new_send(
-            Message::FoundEntry(FoundEntry {
-                data: vec![],
-                misses: vec![PlacementMiss::new(placement_key, node.did())],
-                resource: entry.did,
-                redundancy: 2,
-            }),
-            &context_session,
-            node.did(),
-            node.did(),
-        )?;
-
-        handler
-            .handle(&context, &FoundEntry {
-                data: vec![],
-                misses: vec![PlacementMiss::new(placement_key, node.did())],
-                resource: entry.did,
-                redundancy: 2,
-            })
-            .await?;
-        node.swarm
-            .transport
-            .expire_storage_lookup_observation(entry.did, 2)?;
-        handler
-            .handle(&context, &FoundEntry {
-                data: vec![entry.clone()],
-                misses: vec![],
-                resource: entry.did,
-                redundancy: 2,
-            })
-            .await?;
-
-        assert_eq!(node.swarm.storage_check_cache(entry.did).await, Some(entry));
-        assert_eq!(
-            node.dht().storage.get(&placement_key.to_string()).await?,
-            None
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn sync_entries_handler_reports_persisted_entries() -> Result<()> {
-        let sender = prepare_node(SecretKey::random()).await;
-        let receiver = prepare_node(SecretKey::random()).await;
-        manually_establish_connection(&sender.swarm, &receiver.swarm).await;
-        wait_for_msgs([&sender, &receiver]).await;
-        assert_no_more_msg([&sender, &receiver]).await;
-        for successor in receiver.dht().successors().list()? {
-            receiver.dht().successors().remove(successor)?;
-        }
-        *receiver.dht().lock_predecessor()? = None;
-
-        let receiver_handler =
-            MessageHandler::new(receiver.swarm.transport.clone(), Arc::new(NoopCallback));
-        let placement_key = Did::from(100u32);
-        assert!(matches!(
-            receiver.dht().find_storage_owner(placement_key)?,
-            PeerRingAction::Some(owner) if owner == receiver.did()
-        ));
-        let entry = Entry::new(
-            Did::from(10u32),
-            vec!["handler acked".to_string().encode()?],
-            EntryKind::Data,
-        );
-        let stored_entry = entry.clone().try_into_storage_entry()?;
-        let sync_msg = SyncEntriesWithSuccessor {
-            destination: StorageSyncDestination::PhysicalOwner(receiver.did()),
-            data: vec![PlacedEntry::new(placement_key, entry.clone())],
-        };
-        let context = MessagePayload::new_send(
-            Message::SyncEntriesWithSuccessor(sync_msg.clone()),
-            sender.swarm.transport.session_sk(),
-            receiver.did(),
-            receiver.did(),
-        )?;
-
-        receiver_handler.handle(&context, &sync_msg).await?;
-
-        let payload = next_payload(&sender).await?;
-        match payload.transaction.data::<Message>()? {
-            Message::SyncEntriesWithSuccessorReport(report) => {
-                assert_eq!(report.acks, vec![SyncedEntryAck::new(
-                    placement_key,
-                    stored_entry.clone()
-                )]);
-            }
-            message => {
-                return Err(Error::InvalidMessage(format!(
-                    "expected SyncEntriesWithSuccessorReport, got {message:?}"
-                )))
-            }
-        }
-        assert_eq!(
-            receiver
-                .dht()
-                .storage
-                .get(&placement_key.to_string())
-                .await?,
-            Some(stored_entry)
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn persist_synced_entries_returns_acks_for_owned_entries() -> Result<()> {
-        let receiver = prepare_node(SecretKey::random()).await;
-        let handler = MessageHandler::new(receiver.swarm.transport.clone(), Arc::new(NoopCallback));
-        let placement_key = Did::from(100u32);
-        let entry = Entry::new(
-            Did::from(10u32),
-            vec!["acked".to_string().encode()?],
-            EntryKind::Data,
-        );
-        let stored_entry = entry.clone().try_into_storage_entry()?;
-        let sync_msg = SyncEntriesWithSuccessor {
-            destination: StorageSyncDestination::PhysicalOwner(receiver.did()),
-            data: vec![PlacedEntry::new(placement_key, entry.clone())],
-        };
-
-        let acks = persist_synced_entries(&handler, &sync_msg).await?;
-
-        assert_eq!(acks, vec![SyncedEntryAck::new(
-            placement_key,
-            stored_entry.clone()
-        )]);
-        assert_eq!(
-            receiver
-                .dht()
-                .storage
-                .get(&placement_key.to_string())
-                .await?,
-            Some(stored_entry)
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn sync_entries_handler_skips_entries_owned_by_another_virtual_owner() -> Result<()> {
-        let mut keys = gen_ordered_keys(2).into_iter();
-        let sender = prepare_node_with_virtual_nodes(next_generated_key(&mut keys)?, 2)?;
-        let receiver = prepare_node_with_virtual_nodes(next_generated_key(&mut keys)?, 2)?;
-        manually_establish_connection(&sender.swarm, &receiver.swarm).await;
-        wait_for_msgs([&sender, &receiver]).await;
-        assert_no_more_msg([&sender, &receiver]).await;
-        let _ = receiver.dht().join(sender.did())?;
-
-        let placement_key = receiver
-            .dht()
-            .storage_virtual_positions(sender.did())?
-            .into_iter()
-            .next()
-            .map(|position| position.vnode_did)
-            .ok_or_else(|| Error::InvalidMessage("expected sender virtual position".to_string()))?;
-        assert!(matches!(
-            receiver.dht().find_storage_owner(placement_key)?,
-            PeerRingAction::RemoteAction(owner, PeerRingRemoteAction::FindSuccessor(key))
-                if owner == sender.did() && key == placement_key
-        ));
-
-        let entry = Entry::new(
-            Did::from(10u32),
-            vec!["wrong owner".to_string().encode()?],
-            EntryKind::Data,
-        );
-        let stored_entry = entry.clone().try_into_storage_entry()?;
-        sender
-            .dht()
-            .storage
-            .put(&placement_key.to_string(), &stored_entry)
-            .await?;
-        let sync_msg = SyncEntriesWithSuccessor {
-            destination: StorageSyncDestination::PhysicalOwner(receiver.did()),
-            data: vec![PlacedEntry::new(placement_key, entry)],
-        };
-        let context = MessagePayload::new_send(
-            Message::SyncEntriesWithSuccessor(sync_msg.clone()),
-            sender.swarm.transport.session_sk(),
-            receiver.did(),
-            receiver.did(),
-        )?;
-        sender.swarm.transport.record_pending_storage_sync_ack(
-            context.transaction.tx_id,
-            sync_msg.destination,
-            receiver.did(),
-            &sync_msg.data,
-        )?;
-        let receiver_handler =
-            MessageHandler::new(receiver.swarm.transport.clone(), Arc::new(NoopCallback));
-
-        receiver_handler.handle(&context, &sync_msg).await?;
-
-        assert_eq!(
-            receiver
-                .dht()
-                .storage
-                .get(&placement_key.to_string())
-                .await?,
-            None
-        );
-        let payload = next_payload_for_tx(&sender, context.transaction.tx_id).await?;
-        match payload.transaction.data::<Message>()? {
-            Message::SyncEntriesWithSuccessorReport(report) => {
-                assert!(report.acks.is_empty());
-            }
-            message => {
-                return Err(Error::InvalidMessage(format!(
-                    "expected SyncEntriesWithSuccessorReport, got {message:?}"
-                )))
-            }
-        }
-        assert_eq!(
-            sender.dht().storage.get(&placement_key.to_string()).await?,
-            Some(stored_entry)
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn sync_entries_physical_destination_routes_by_physical_did_not_storage_owner(
-    ) -> Result<()> {
-        let mut keys = gen_ordered_keys(6).into_iter();
-        let node = prepare_node_with_virtual_nodes(next_generated_key(&mut keys)?, 4)?;
-        let mut peers = Vec::new();
-        for _ in 0..5 {
-            peers.push(next_generated_key(&mut keys)?.address().into());
-        }
-        for peer in peers.iter().copied() {
-            let _ = node.dht().join(peer)?;
-        }
-
-        let dht = node.dht();
-        let mut witness = None;
-        for destination in peers {
-            let physical_next = physical_sync_route_next_hop(&dht, destination)?;
-            let storage_next = storage_sync_route_next_hop(&dht, destination)?;
-            if physical_next != storage_next {
-                witness = Some((destination, physical_next, storage_next));
-                break;
-            }
-        }
-        let Some((destination, physical_next, storage_next)) = witness else {
-            return Err(Error::InvalidMessage(
-                "expected physical and storage routes to diverge".to_string(),
-            ));
-        };
-
-        let msg = SyncEntriesWithSuccessor {
-            destination: StorageSyncDestination::PhysicalOwner(destination),
-            data: vec![],
-        };
-        let context = MessagePayload::new_send(
-            Message::SyncEntriesWithSuccessor(msg.clone()),
-            node.swarm.transport.session_sk(),
-            node.did(),
-            destination,
-        )?;
-        let handler = MessageHandler::new(node.swarm.transport.clone(), Arc::new(NoopCallback));
-
-        let next = next_hop_for_sync_entries(&handler, &context, &msg)?;
-
-        assert_eq!(next, physical_next);
-        assert_ne!(next, storage_next);
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn sync_entries_report_handler_deletes_only_acked_keys() -> Result<()> {
-        let node = prepare_node(SecretKey::random()).await;
-        let handler = MessageHandler::new(node.swarm.transport.clone(), Arc::new(NoopCallback));
-        let acked_key = Did::from(100u32);
-        let pending_key = Did::from(120u32);
-        let acked_entry = Entry::new(Did::from(10u32), vec![], EntryKind::Data);
-        let acked_storage_entry = acked_entry.clone().try_into_storage_entry()?;
-        let pending_entry = Entry::new(Did::from(20u32), vec![], EntryKind::Data);
-        let sync_msg = SyncEntriesWithSuccessor {
-            destination: StorageSyncDestination::PhysicalOwner(node.did()),
-            data: vec![PlacedEntry::new(acked_key, acked_entry.clone())],
-        };
-        let request = MessagePayload::new_send(
-            Message::SyncEntriesWithSuccessor(sync_msg.clone()),
-            node.swarm.transport.session_sk(),
-            node.did(),
-            node.did(),
-        )?;
-        node.swarm.transport.record_pending_storage_sync_ack(
-            request.transaction.tx_id,
-            sync_msg.destination,
-            node.did(),
-            &sync_msg.data,
-        )?;
-        node.dht()
-            .storage
-            .put(&acked_key.to_string(), &acked_entry)
-            .await?;
-        node.dht()
-            .storage
-            .put(&pending_key.to_string(), &pending_entry)
-            .await?;
-
-        let report = SyncEntriesWithSuccessorReport::new(sync_msg.destination, node.did(), vec![
-            SyncedEntryAck::new(acked_key, acked_storage_entry),
-        ]);
-        let context = storage_sync_report_payload(
-            &request,
-            report.clone(),
-            node.swarm.transport.session_sk(),
-            node.did(),
-            node.did(),
-        )?;
-
-        handler.handle(&context, &report).await?;
-
-        assert_eq!(node.dht().storage.get(&acked_key.to_string()).await?, None);
-        assert_eq!(
-            node.dht().storage.get(&pending_key.to_string()).await?,
-            Some(pending_entry)
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn sync_entries_report_handler_rejects_untracked_acks() -> Result<()> {
-        let node = prepare_node(SecretKey::random()).await;
-        let handler = MessageHandler::new(node.swarm.transport.clone(), Arc::new(NoopCallback));
-        let acked_key = Did::from(100u32);
-        let acked_entry = Entry::new(Did::from(10u32), vec![], EntryKind::Data);
-        let acked_storage_entry = acked_entry.clone().try_into_storage_entry()?;
-        node.dht()
-            .storage
-            .put(&acked_key.to_string(), &acked_entry)
-            .await?;
-
-        let sync_msg = SyncEntriesWithSuccessor {
-            destination: StorageSyncDestination::PhysicalOwner(node.did()),
-            data: vec![PlacedEntry::new(acked_key, acked_entry.clone())],
-        };
-        let request = MessagePayload::new_send(
-            Message::SyncEntriesWithSuccessor(sync_msg.clone()),
-            node.swarm.transport.session_sk(),
-            node.did(),
-            node.did(),
-        )?;
-        let report = SyncEntriesWithSuccessorReport::new(sync_msg.destination, node.did(), vec![
-            SyncedEntryAck::new(acked_key, acked_storage_entry),
-        ]);
-        let context = storage_sync_report_payload(
-            &request,
-            report.clone(),
-            node.swarm.transport.session_sk(),
-            node.did(),
-            node.did(),
-        )?;
-
-        let result = handler.handle(&context, &report).await;
-
-        assert!(matches!(
-            result,
-            Err(Error::InvalidMessage(message)) if message.contains("no pending capability")
-        ));
-        assert_eq!(
-            node.dht().storage.get(&acked_key.to_string()).await?,
-            Some(acked_entry)
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn sync_entries_report_handler_forwards_before_pending_capability_check() -> Result<()> {
-        let sender = prepare_node(SecretKey::random()).await;
-        let relay = prepare_node(SecretKey::random()).await;
-        let receiver = prepare_node(SecretKey::random()).await;
-        manually_establish_connection(&sender.swarm, &relay.swarm).await;
-        wait_for_msgs([&sender, &relay]).await;
-        assert_no_more_msg([&sender, &relay]).await;
-
-        let handler = MessageHandler::new(relay.swarm.transport.clone(), Arc::new(NoopCallback));
-        let sync_msg = SyncEntriesWithSuccessor {
-            destination: StorageSyncDestination::PhysicalOwner(receiver.did()),
-            data: vec![],
-        };
-        let request = MessagePayload::new_send(
-            Message::SyncEntriesWithSuccessor(sync_msg.clone()),
-            sender.swarm.transport.session_sk(),
-            receiver.did(),
-            receiver.did(),
-        )?;
-        let report =
-            SyncEntriesWithSuccessorReport::new(sync_msg.destination, receiver.did(), vec![]);
-        let context = storage_sync_report_payload(
-            &request,
-            report.clone(),
-            receiver.swarm.transport.session_sk(),
-            relay.did(),
-            sender.did(),
-        )?;
-
-        handler.handle(&context, &report).await?;
-
-        let forwarded = next_payload_for_tx(&sender, request.transaction.tx_id).await?;
-        match forwarded.transaction.data::<Message>()? {
-            Message::SyncEntriesWithSuccessorReport(forwarded_report) => {
-                assert_eq!(forwarded_report.destination, report.destination);
-                assert_eq!(forwarded_report.receiver, report.receiver);
-                assert_eq!(forwarded_report.acks, report.acks);
-            }
-            message => {
-                return Err(Error::InvalidMessage(format!(
-                    "expected SyncEntriesWithSuccessorReport, got {message:?}"
-                )))
-            }
-        }
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn sync_entries_report_handler_rejects_wrong_physical_receiver() -> Result<()> {
-        let sender = prepare_node(SecretKey::random()).await;
-        let receiver = prepare_node(SecretKey::random()).await;
-        let handler = MessageHandler::new(sender.swarm.transport.clone(), Arc::new(NoopCallback));
-        let acked_key = Did::from(100u32);
-        let acked_entry = Entry::new(Did::from(10u32), vec![], EntryKind::Data);
-        let acked_storage_entry = acked_entry.clone().try_into_storage_entry()?;
-        sender
-            .dht()
-            .storage
-            .put(&acked_key.to_string(), &acked_entry)
-            .await?;
-
-        let sync_msg = SyncEntriesWithSuccessor {
-            destination: StorageSyncDestination::PhysicalOwner(receiver.did()),
-            data: vec![PlacedEntry::new(acked_key, acked_entry.clone())],
-        };
-        let request = MessagePayload::new_send(
-            Message::SyncEntriesWithSuccessor(sync_msg.clone()),
-            sender.swarm.transport.session_sk(),
-            receiver.did(),
-            receiver.did(),
-        )?;
-        sender.swarm.transport.record_pending_storage_sync_ack(
-            request.transaction.tx_id,
-            sync_msg.destination,
-            receiver.did(),
-            &sync_msg.data,
-        )?;
-        let report = SyncEntriesWithSuccessorReport::new(sync_msg.destination, sender.did(), vec![
-            SyncedEntryAck::new(acked_key, acked_storage_entry),
-        ]);
-        let context = storage_sync_report_payload(
-            &request,
-            report.clone(),
-            sender.swarm.transport.session_sk(),
-            sender.did(),
-            sender.did(),
-        )?;
-
-        let result = handler.handle(&context, &report).await;
-
-        assert!(matches!(
-            result,
-            Err(Error::InvalidMessage(message)) if message.contains("receiver does not match pending sync")
-        ));
-        assert_eq!(
-            sender.dht().storage.get(&acked_key.to_string()).await?,
-            Some(acked_entry)
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn sync_entries_report_handler_rejects_unproven_placement_receiver() -> Result<()> {
-        let sender = prepare_node(SecretKey::random()).await;
-        let route_next_hop = prepare_node(SecretKey::random()).await;
-        let final_receiver = prepare_node(SecretKey::random()).await;
-        let handler = MessageHandler::new(sender.swarm.transport.clone(), Arc::new(NoopCallback));
-        let placement_key = Did::from(100u32);
-        let acked_entry = Entry::new(Did::from(10u32), vec![], EntryKind::Data);
-        let acked_storage_entry = acked_entry.clone().try_into_storage_entry()?;
-        sender
-            .dht()
-            .storage
-            .put(&placement_key.to_string(), &acked_entry)
-            .await?;
-
-        let sync_msg = SyncEntriesWithSuccessor {
-            destination: StorageSyncDestination::PlacementKey(placement_key),
-            data: vec![PlacedEntry::new(placement_key, acked_entry.clone())],
-        };
-        let request = MessagePayload::new_send(
-            Message::SyncEntriesWithSuccessor(sync_msg.clone()),
-            sender.swarm.transport.session_sk(),
-            route_next_hop.did(),
-            sync_msg.destination.did(),
-        )?;
-        sender.swarm.transport.record_pending_storage_sync_ack(
-            request.transaction.tx_id,
-            sync_msg.destination,
-            route_next_hop.did(),
-            &sync_msg.data,
-        )?;
-        let report =
-            SyncEntriesWithSuccessorReport::new(sync_msg.destination, final_receiver.did(), vec![
-                SyncedEntryAck::new(placement_key, acked_storage_entry),
-            ]);
-        let context = storage_sync_report_payload(
-            &request,
-            report.clone(),
-            final_receiver.swarm.transport.session_sk(),
-            sender.did(),
-            sender.did(),
-        )?;
-
-        let result = handler.handle(&context, &report).await;
-
-        assert!(matches!(
-            result,
-            Err(Error::InvalidMessage(message)) if message.contains("receiver does not match pending sync")
-        ));
-        assert_eq!(
-            sender.dht().storage.get(&placement_key.to_string()).await?,
-            Some(acked_entry)
-        );
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn storage_store_fetches_remote_entry_into_cache() -> Result<()> {
-        let mut keys = gen_ordered_keys(2).into_iter();
-        let key1 = next_generated_key(&mut keys)?;
-        let key2 = next_generated_key(&mut keys)?;
-        let node1 = prepare_node(key1).await;
-        let node2 = prepare_node(key2).await;
-
-        manually_establish_connection(&node1.swarm, &node2.swarm).await;
-        wait_for_msgs([&node1, &node2]).await;
-        assert_no_more_msg([&node1, &node2]).await;
-
-        // Now, node1 is the successor of node2, and node2 is the successor of node1.
-        // Following tests storing data on node2 and query it from node1.
-        let data = "Across the Great Wall we can reach every corner in the world.".to_string();
-        let entry: Entry = data.clone().try_into()?;
-        let entry_key = entry.did;
-
-        // Make sure the data is stored on node2.
-        let (node1, node2) = if entry_key.in_range(node2.did(), node2.did(), node1.did()) {
-            (node1, node2)
-        } else {
-            (node2, node1)
-        };
-
-        assert_eq!(node1.dht().cache.count().await?, 0);
-        assert_eq!(node2.dht().cache.count().await?, 0);
-        assert!(node1.swarm.storage_check_cache(entry_key).await.is_none());
-        assert!(node2.swarm.storage_check_cache(entry_key).await.is_none());
-
-        <Swarm as ChordStorageInterface<1>>::storage_store(&node1.swarm, entry.clone()).await?;
-        let ev = next_payload(&node2).await?;
-        assert!(matches!(
-            ev.transaction.data()?,
-            Message::OperateEntry(PlacedEntryOperation {
-                placement,
-                op: EntryOperation::Overwrite(x),
-            }) if placement == entry_key && x.did == entry_key
-        ));
-
-        assert!(node1.swarm.storage_check_cache(entry_key).await.is_none());
-        assert!(node2.swarm.storage_check_cache(entry_key).await.is_none());
-        assert!(node1.dht().storage.count().await? == 0);
-        assert!(node2.dht().storage.count().await? != 0);
-
-        // test remote query
-        println!("entry_key is on node2 {:?}", node2.did());
-        <Swarm as ChordStorageInterface<1>>::storage_fetch(&node1.swarm, entry_key).await?;
-
-        // it will send request to node2
-        let ev = next_payload(&node2).await?;
-        // node2 received search entry request
-        assert!(matches!(
-            ev.transaction.data()?,
-            Message::SearchEntry(x) if x.resource == entry_key && x.placement == entry_key
-        ));
-
-        let ev = next_payload(&node1).await?;
-        assert!(matches!(
-            ev.transaction.data()?,
-            Message::FoundEntry(x)
-                if x.resource == entry_key
-                    && x.misses.is_empty()
-                    && x.data.first().is_some_and(|entry| entry.did == entry_key)
-        ));
-
-        assert_cached_data_values(&node1, entry_key, &[data.as_str()]).await?;
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn storage_append_data_preserves_entry_payload_order() -> Result<()> {
-        let mut keys = gen_ordered_keys(2).into_iter();
-        let key1 = next_generated_key(&mut keys)?;
-        let key2 = next_generated_key(&mut keys)?;
-        let node1 = prepare_node(key1).await;
-        let node2 = prepare_node(key2).await;
-
-        manually_establish_connection(&node1.swarm, &node2.swarm).await;
-        wait_for_msgs([&node1, &node2]).await;
-        assert_no_more_msg([&node1, &node2]).await;
-
-        // Now, node1 is the successor of node2, and node2 is the successor of node1.
-        // Following tests storing data on node2 and query it from node1.
-        let topic = "Across the Great Wall we can reach every corner in the world.".to_string();
-        let entry: Entry = topic.clone().try_into()?;
-        let entry_key = entry.did;
-
-        // Make sure the data is stored on node2.
-        let (node1, node2) = if entry_key.in_range(node2.did(), node2.did(), node1.did()) {
-            (node1, node2)
-        } else {
-            (node2, node1)
-        };
-
-        assert_eq!(node1.dht().cache.count().await?, 0);
-        assert_eq!(node2.dht().cache.count().await?, 0);
-        assert!(node1.swarm.storage_check_cache(entry_key).await.is_none());
-        assert!(node2.swarm.storage_check_cache(entry_key).await.is_none());
-
-        <Swarm as ChordStorageInterface<1>>::storage_append_data(
-            &node1.swarm,
-            &topic,
-            "111".to_string().encode()?,
-        )
-        .await?;
-        wait_for_msgs([&node1, &node2]).await;
-        assert_no_more_msg([&node1, &node2]).await;
-
-        <Swarm as ChordStorageInterface<1>>::storage_append_data(
-            &node1.swarm,
-            &topic,
-            "222".to_string().encode()?,
-        )
-        .await?;
-        wait_for_msgs([&node1, &node2]).await;
-        assert_no_more_msg([&node1, &node2]).await;
-
-        assert!(node1.swarm.storage_check_cache(entry_key).await.is_none());
-        assert!(node2.swarm.storage_check_cache(entry_key).await.is_none());
-        assert!(node1.dht().storage.count().await? == 0);
-        assert!(node2.dht().storage.count().await? != 0);
-
-        // test remote query
-        println!("entry_key is on node2 {:?}", node2.did());
-        <Swarm as ChordStorageInterface<1>>::storage_fetch(&node1.swarm, entry_key).await?;
-        wait_for_msgs([&node1, &node2]).await;
-        assert_no_more_msg([&node1, &node2]).await;
-
-        assert_cached_data_values(&node1, entry_key, &["111", "222"]).await?;
-
-        // Append more data
-        <Swarm as ChordStorageInterface<1>>::storage_append_data(
-            &node1.swarm,
-            &topic,
-            "333".to_string().encode()?,
-        )
-        .await?;
-        wait_for_msgs([&node1, &node2]).await;
-        assert_no_more_msg([&node1, &node2]).await;
-
-        // test remote query agagin
-        println!("entry_key is on node2 {:?}", node2.did());
-        <Swarm as ChordStorageInterface<1>>::storage_fetch(&node1.swarm, entry_key).await?;
-        wait_for_msgs([&node1, &node2]).await;
-        assert_no_more_msg([&node1, &node2]).await;
-
-        assert_cached_data_values(&node1, entry_key, &["111", "222", "333"]).await?;
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn storage_touch_data_moves_existing_entry_payload_to_end_once() -> Result<()> {
-        let mut keys = gen_ordered_keys(2).into_iter();
-        let key1 = next_generated_key(&mut keys)?;
-        let key2 = next_generated_key(&mut keys)?;
-        let node1 = prepare_node(key1).await;
-        let node2 = prepare_node(key2).await;
-
-        manually_establish_connection(&node1.swarm, &node2.swarm).await;
-        wait_for_msgs([&node1, &node2]).await;
-        assert_no_more_msg([&node1, &node2]).await;
-
-        let topic = "touch keeps unique entry payloads ordered by recency".to_string();
-        let entry: Entry = topic.clone().try_into()?;
-        let entry_key = entry.did;
-
-        let (node1, node2) = if entry_key.in_range(node2.did(), node2.did(), node1.did()) {
-            (node1, node2)
-        } else {
-            (node2, node1)
-        };
-
-        for value in ["111", "222", "333", "222"] {
-            <Swarm as ChordStorageInterface<1>>::storage_touch_data(
-                &node1.swarm,
-                &topic,
-                value.to_string().encode()?,
-            )
-            .await?;
-            wait_for_msgs([&node1, &node2]).await;
-            assert_no_more_msg([&node1, &node2]).await;
-        }
-
-        assert!(node1.swarm.storage_check_cache(entry_key).await.is_none());
-        assert!(node2.swarm.storage_check_cache(entry_key).await.is_none());
-        assert_eq!(node1.dht().storage.count().await?, 0);
-        assert_ne!(node2.dht().storage.count().await?, 0);
-
-        <Swarm as ChordStorageInterface<1>>::storage_fetch(&node1.swarm, entry_key).await?;
-        wait_for_msgs([&node1, &node2]).await;
-        assert_no_more_msg([&node1, &node2]).await;
-
-        assert_cached_data_values(&node1, entry_key, &["111", "333", "222"]).await?;
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn storage_tombstone_data_removes_observed_payload() -> Result<()> {
-        let mut keys = gen_ordered_keys(2).into_iter();
-        let key1 = next_generated_key(&mut keys)?;
-        let key2 = next_generated_key(&mut keys)?;
-        let node1 = prepare_node(key1).await;
-        let node2 = prepare_node(key2).await;
-
-        manually_establish_connection(&node1.swarm, &node2.swarm).await;
-        wait_for_msgs([&node1, &node2]).await;
-        assert_no_more_msg([&node1, &node2]).await;
-
-        let topic = "tombstone removes stale data topic payloads".to_string();
-        let entry: Entry = topic.clone().try_into()?;
-        let entry_key = entry.did;
-
-        let (node1, node2) = if entry_key.in_range(node2.did(), node2.did(), node1.did()) {
-            (node1, node2)
-        } else {
-            (node2, node1)
-        };
-
-        for value in ["111", "222"] {
-            <Swarm as ChordStorageInterface<1>>::storage_touch_data(
-                &node1.swarm,
-                &topic,
-                value.to_string().encode()?,
-            )
-            .await?;
-            wait_for_msgs([&node1, &node2]).await;
-            assert_no_more_msg([&node1, &node2]).await;
-        }
-
-        <Swarm as ChordStorageInterface<1>>::storage_tombstone_data(
-            &node1.swarm,
-            &topic,
-            "111".to_string().encode()?,
-        )
-        .await?;
-        wait_for_msgs([&node1, &node2]).await;
-        assert_no_more_msg([&node1, &node2]).await;
-
-        <Swarm as ChordStorageInterface<1>>::storage_fetch(&node1.swarm, entry_key).await?;
-        wait_for_msgs([&node1, &node2]).await;
-        assert_no_more_msg([&node1, &node2]).await;
-
-        assert_cached_data_values(&node1, entry_key, &["222"]).await?;
-
-        Ok(())
-    }
-}
+mod api_tests;
+#[cfg(not(feature = "wasm"))]
+#[cfg(test)]
+mod repair_tests;
+#[cfg(not(feature = "wasm"))]
+#[cfg(test)]
+mod sync_ack_tests;
+#[cfg(not(feature = "wasm"))]
+#[cfg(test)]
+mod sync_receive_tests;
+#[cfg(not(feature = "wasm"))]
+#[cfg(test)]
+mod sync_report_tests;
+#[cfg(not(feature = "wasm"))]
+#[cfg(test)]
+mod test_support;
